requirements.txt
fastapi
uvicorn
apify-client
openai
python-dotenv
requests
celery[redis]
redis
flower